backend/requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.2
gunicorn>=22.0
djangorestframework>=3.15
django-cors-headers>=4.4
django-localflavor>=4.0